use realpath if installed, xargs readlink does not work on all ubuntu versions

+++ code/bin/scrubStorage.sh
@@ -139,11 +139,15 @@ sweep () {
    done
 
    # next remove all archive data that do not have a reference in tmp anymore
-   ls -d /data/scratch/tmp.*/INPUT | xargs readlink -f | sort | uniq > /tmp/tmpreferenced
-   ls -d /data/scratch/archive/* > /tmp/inarchive
-   notreferenced=`grep -v -f /tmp/tmpreferenced /tmp/inarchive`
-   echo "REMOVE: archive data not referenced in /data/scratch/tmp.* anymore \"$notreferenced\"" >> $log
-   grep -v -f /tmp/tmpreferenced /tmp/inarchive | xargs sudo \rm -R -f 
+   if hash realpath 2>/dev/null; then
+     ls -d /data/scratch/tmp.*/INPUT | xargs realpath | sort | uniq > /tmp/tmpreferenced
+     ls -d /data/scratch/archive/* > /tmp/inarchive
+     notreferenced=`grep -v -f /tmp/tmpreferenced /tmp/inarchive`
+     echo "REMOVE: archive data not referenced in /data/scratch/tmp.* anymore \"$notreferenced\"" >> $log
+     grep -v -f /tmp/tmpreferenced /tmp/inarchive | xargs sudo \rm -R -f 
+   else
+     echo "ERROR: could not remove non-referenced scans because realpath is not installed" >> $log
+   fi
 
    # next remove all docker container that are un-named and not running
    old=`sudo docker ps -a | grep Exit | awk '{print $1}'`
